+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/MaintenanceProxyService.java
@@ -33,6 +33,7 @@ public class MaintenanceProxyService extends StatelessService {
     private Consumer<Operation> parentHandler;
 
     public MaintenanceProxyService(Consumer<Operation> parentHandler) {
+        super.toggleOption(ServiceOption.CORE, true);
         super.toggleOption(ServiceOption.PERIODIC_MAINTENANCE, true);
         this.parentHandler = parentHandler;
     }